added response parsing
