+++ tutorials/01_Basic-Usecase/01_Basic-Usecase.toml
@@ -1,5 +1,4 @@
 [METADATA]
-
 version = 0.1
 description = "A sample prompt for RadPrompter"
