+++ tutorials/02_RDP-Templating/02_RDP-Templating.toml
@@ -1,5 +1,4 @@
 [METADATA]
-
 version = 0.1
 description = "A sample prompt for RadPrompter"
